backend/requirements.txt
# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-multipart>=0.0.9

# Google GenAI SDK (official SDK for Gemini)
google-genai>=1.0.0

# Image/Video Processing
mediapipe>=0.10.9
opencv-python-headless>=4.9.0
pillow>=10.2.0
numpy>=1.26.0

# WebSockets
websockets>=12.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0
